refactor: Move `active-session` into user sessions
