+++ frontend/src/app/app.module.ts
@@ -86,13 +86,13 @@ import { ProjectUserSettingsComponent } from './projects/project-detail/project-
 import { ProjectOverviewComponent } from './projects/project-overview/project-overview.component';
 import { ProjectWrapperComponent } from './projects/project-wrapper/project-wrapper.component';
 import { WhitespaceUrlInterceptor } from './services/encoder/encoder.interceptor';
-import { ActiveSessionsComponent } from './sessions/active-sessions/active-sessions.component';
-import { FileBrowserComponent } from './sessions/active-sessions/file-browser/file-browser.component';
-import { FileExistsDialogComponent } from './sessions/active-sessions/file-browser/file-exists-dialog/file-exists-dialog.component';
 import { DeleteSessionDialogComponent } from './sessions/delete-session-dialog/delete-session-dialog.component';
 import { GuacamoleComponent } from './sessions/session-created/guacamole/guacamole.component';
 import { SessionOverviewComponent } from './sessions/session-overview/session-overview.component';
 import { SessionsComponent } from './sessions/sessions.component';
+import { ActiveSessionsComponent } from './sessions/user-sessions-wrapper/active-sessions/active-sessions.component';
+import { FileBrowserComponent } from './sessions/user-sessions-wrapper/active-sessions/file-browser/file-browser.component';
+import { FileExistsDialogComponent } from './sessions/user-sessions-wrapper/active-sessions/file-browser/file-exists-dialog/file-exists-dialog.component';
 import { CreatePersistentSessionComponent } from './sessions/user-sessions-wrapper/create-sessions/create-persistent-session/create-persistent-session.component';
 import { CreateReadonlyModelOptionsComponent } from './sessions/user-sessions-wrapper/create-sessions/create-readonly-session/create-readonly-model-options/create-readonly-model-options.component';
 import { CreateReadonlySessionDialogComponent } from './sessions/user-sessions-wrapper/create-sessions/create-readonly-session/create-readonly-session-dialog.component';
