+++ frontend/src/app/sessions/user-sessions-wrapper/active-sessions/active-sessions.component.css
