+++ frontend/src/app/sessions/user-sessions-wrapper/active-sessions/file-browser/file-browser.component.html
